# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NexGuardian Foundation API",
    "version": "0.1.0",
    "description": "Contrato proposto, sem backend. Nenhum endpoint do produto original é utilizado."
  },
  "servers": [
    {
      "url": "https://api.nexguardian.invalid/v1",
      "description": "Placeholder não operacional; substituir por infraestrutura própria."
    }
  ],
...
  "paths": {
...
    "/devices/pair": {
      "post": {
        "operationId": "requestPairing",
        "tags": [
          "Devices"
        ],
        "summary": "Solicitar vínculo",
        "description": "Solicitar vínculo",
        "security": [
          {
            "activationTicket": []
          }
        ],
        "responses": {
          "201": {
            "description": "Sucesso",
            "headers": {
              "Cache-Control": {
                "description": "no-store",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairResponse"
                }
              }
            }
          },
          "400": {
            "description": "Pedido inválido",
            "headers": {
              "Cache-Control": {
                "description": "no-store",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Credencial inválida ou expirada",
            "headers": {
              "Cache-Control": {
                "description": "no-store",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Acesso não autorizado",
            "headers": {
              "Cache-Control": {
                "description": "no-store",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflito de estado ou repetição divergente",
            "headers": {
              "Cache-Control": {
                "description": "no-store",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Limite de tentativas excedido",
            "headers": {
              "Cache-Control": {
                "description": "no-store",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno sem detalhes sensíveis",
            "headers": {
              "Cache-Control": {
                "description": "no-store",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PairRequest"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Escopo por principal/operação. Mesmo corpo retorna mesmo resultado; corpo diferente retorna 409."
          }
        ]
      }
    },
    "/devices/pair/confirm": {
      "post": {
        "operationId": "confirmPairing",
        "tags": [
          "Devices"
        ],
        "summary": "Confirmar vínculo e obter estado de assinatura",
        "description": "Confirmar vínculo e obter estado de assinatura",
        "security": [
          {
            "pairingTicket": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "headers": {
              "Cache-Control": {
                "description": "no-store",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmPairResponse"
                }
              }
            }
          },
          "400": {
            "description": "Pedido inválido",
            "headers": {
              "Cache-Control": {
                "description": "no-store",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Credencial inválida ou expirada",
            "headers": {
              "Cache-Control": {
                "description": "no-store",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Acesso não autorizado",
            "headers": {
              "Cache-Control": {
                "description": "no-store",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflito de estado ou repetição divergente",
            "headers": {
              "Cache-Control": {
                "description": "no-store",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Limite de tentativas excedido",
            "headers": {
              "Cache-Control": {
                "description": "no-store",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno sem detalhes sensíveis",
            "headers": {
              "Cache-Control": {
                "description": "no-store",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmPairRequest"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Escopo por principal/operação. Mesmo corpo retorna mesmo resultado; corpo diferente retorna 409."
          }
        ]
      }
    },
    "/devices/session/refresh": {
      "post": {
        "operationId": "refreshDeviceSession",
        "tags": [
          "Devices"
        ],
        "summary": "Rotacionar sessão de dispositivo",
        "description": "Refresh token autentica a operação. Rotação atômica, dispositivo vinculado e família de sessões; replay invalida a família. Nunca aceitar UUID sozinho.",
        "security": [],
        "responses": {
          "200": {
            "description": "Sucesso",
            "headers": {
              "Cache-Control": {
                "description": "no-store",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceSession"
                }
              }
            }
          },
          "400": {
            "description": "Pedido inválido",
            "headers": {
              "Cache-Control": {
                "description": "no-store",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Credencial inválida ou expirada",
            "headers": {
              "Cache-Control": {
                "description": "no-store",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Acesso não autorizado",
            "headers": {
              "Cache-Control": {
                "description": "no-store",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflito de estado ou repetição divergente",
            "headers": {
              "Cache-Control": {
                "description": "no-store",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Limite de tentativas excedido",
            "headers": {
              "Cache-Control": {
                "description": "no-store",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno sem detalhes sensíveis",
            "headers": {
              "Cache-Control": {
                "description": "no-store",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          }
        }
      }
    },
    "/devices/heartbeat": {
      "post": {
        "operationId": "heartbeat",
        "tags": [
          "Devices"
        ],
        "summary": "Enviar status operacional mínimo",
        "description": "Enviar status operacional mínimo",
        "security": [
          {
            "deviceBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "headers": {
              "Cache-Control": {
                "description": "no-store",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HeartbeatResponse"
                }
              }
            }
          },
          "400": {
            "description": "Pedido inválido",
            "headers": {
              "Cache-Control": {
                "description": "no-store",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Credencial inválida ou expirada",
            "headers": {
              "Cache-Control": {
                "description": "no-store",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Acesso não autorizado",
            "headers": {
              "Cache-Control": {
                "description": "no-store",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflito de estado ou repetição divergente",
            "headers": {
              "Cache-Control": {
                "description": "no-store",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Limite de tentativas excedido",
            "headers": {
              "Cache-Control": {
                "description": "no-store",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno sem detalhes sensíveis",
            "headers": {
              "Cache-Control": {
                "description": "no-store",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HeartbeatRequest"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Escopo por principal/operação. Mesmo corpo retorna mesmo resultado; corpo diferente retorna 409."
          }
        ]
      }
    },
    "/devices/revoke": {
      "post": {
        "operationId": "revokeDevice",
        "tags": [
          "Devices"
        ],
        "summary": "Revogar vínculo",
        "description": "Dispositivo só revoga a si; conta só revoga seus dispositivos. Token revogado não autoriza novas operações. Repetição por conta com mesma chave mantém resultado.",
        "security": [
          {
            "deviceBearer": []
          },
          {
            "accountBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "headers": {
              "Cache-Control": {
                "description": "no-store",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevokeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Pedido inválido",
            "headers": {
              "Cache-Control": {
                "description": "no-store",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Credencial inválida ou expirada",
            "headers": {
              "Cache-Control": {
                "description": "no-store",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Acesso não autorizado",
            "headers": {
              "Cache-Control": {
                "description": "no-store",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflito de estado ou repetição divergente",
            "headers": {
              "Cache-Control": {
                "description": "no-store",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Limite de tentativas excedido",
            "headers": {
              "Cache-Control": {
                "description": "no-store",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno sem detalhes sensíveis",
            "headers": {
              "Cache-Control": {
                "description": "no-store",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeRequest"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Escopo por principal/operação. Mesmo corpo retorna mesmo resultado; corpo diferente retorna 409."
          }
        ]
      }
    },
    "/devices/me": {
      "get": {
        "operationId": "getDevice",
        "tags": [
          "Devices"
        ],
        "summary": "Consultar o próprio aparelho",
        "description": "Consultar o próprio aparelho",
        "security": [
          {
            "deviceBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "headers": {
              "Cache-Control": {
                "description": "no-store",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceMeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Pedido inválido",
            "headers": {
              "Cache-Control": {
                "description": "no-store",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Credencial inválida ou expirada",
            "headers": {
              "Cache-Control": {
                "description": "no-store",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Acesso não autorizado",
            "headers": {
              "Cache-Control": {
                "description": "no-store",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflito de estado ou repetição divergente",
            "headers": {
              "Cache-Control": {
                "description": "no-store",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Limite de tentativas excedido",
            "headers": {
              "Cache-Control": {
                "description": "no-store",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno sem detalhes sensíveis",
            "headers": {
              "Cache-Control": {
                "description": "no-store",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/devices": {
      "get": {
        "operationId": "listDevices",
        "tags": [
          "Devices"
        ],
        "summary": "Lista dispositivos do workspace",
        "description": "Escopo de conta pelo token.",
        "security": [
          {
            "accountBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "headers": {
              "Cache-Control": {
                "description": "no-store",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Pedido inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Proibido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflito",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Muitas requisições",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "INVALID_REQUEST",
              "INVALID_CREDENTIALS",
              "ACTIVATION_INVALID",
              "ACTIVATION_EXPIRED",
              "ACTIVATION_CONSUMED",
              "PAIRING_EXPIRED",
              "PAIRING_STATE_CONFLICT",
              "SESSION_REVOKED",
              "REFRESH_REUSED",
              "IDEMPOTENCY_CONFLICT",
              "FORBIDDEN",
              "NOT_FOUND",
              "RATE_LIMITED",
              "SERVER_ERROR"
            ]
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "code",
          "message",
          "requestId"
        ]
      },
...
      "PairRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "deviceId": {
            "type": "string",
            "format": "uuid"
          },
          "deviceInfo": {
            "$ref": "#/components/schemas/DeviceInfo"
          }
        },
        "required": [
          "deviceId",
          "deviceInfo"
        ]
      },
      "PairResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "pairingId": {
            "type": "string",
            "format": "uuid"
          },
          "device": {
            "$ref": "#/components/schemas/Device"
          },
          "pairingTicket": {
            "type": "string",
            "minLength": 32
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "description": "RFC3339 UTC; datas de negócio são definidas pelo servidor."
          },
          "serverTime": {
            "type": "string",
            "format": "date-time",
            "description": "RFC3339 UTC; datas de negócio são definidas pelo servidor."
          }
        },
        "required": [
          "pairingId",
          "device",
          "pairingTicket",
          "expiresAt",
          "serverTime"
        ],
        "description": "Estado PAIRING_PENDING. Ticket só confirma este pairingId/deviceId; não autoriza APIs de dispositivo."
      },
...
      "ConfirmPairResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "device": {
            "$ref": "#/components/schemas/Device"
          },
          "session": {
            "$ref": "#/components/schemas/DeviceSession"
          },
          "subscription": {
            "$ref": "#/components/schemas/Subscription"
          },
          "serverTime": {
            "type": "string",
            "format": "date-time",
            "description": "RFC3339 UTC; datas de negócio são definidas pelo servidor."
          }
        },
        "required": [
          "device",
          "session",
          "subscription",
          "serverTime"
        ],
        "description": "Confirmação atômica: consome código/ticket, marca PAIRED, cria sessão e inicia trial somente se a conta nunca teve trial. confirmedOnDevice documenta a confirmação de UI, não é prova criptográfica."
      },
...
      "RevokeRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "deviceId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "deviceId"
        ]
      },
...
    },
    "securitySchemes": {
      "accountBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Access token de conta. O escopo de isolamento (accountId, que mapeia o Workspace do domínio — ver ADR-0003) é transportado como claim no token; nenhum endpoint aceita identificador de conta/workspace por query ou path."
      },
      "deviceBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Access token de sessão de dispositivo, com audiência de dispositivo, vinculado a deviceId e ao escopo de conta/workspace (accountId) transportado como claim no token; nenhum endpoint aceita esse identificador por query ou path."
      },
      "activationTicket": {
        "type": "http",
        "scheme": "bearer",
        "description": "Ticket temporário exclusivo de validate → pair, ligado à conta e ao UUID da instalação."
      },
      "pairingTicket": {
        "type": "http",
        "scheme": "bearer",
        "description": "Ticket temporário exclusivo para confirmar um pareamento pendente."
      }
    }
  }
}